combine empty and nonempty dashboards
