+++ src/routes.ts
@@ -11,10 +11,8 @@ import Dashboard from "./views/Dashboard";
 import Profile from "./views/Profile";
 import ReviewOpportunity from "./views/ReviewOpportunity";
 import SignUpConfirm from "./views/SignUpConfirm";
-import OrganizationDashboard from "./views/OrganizationDashboard";
 
 export const routes = [
-    // { path: "/dashboard", exact: true, component: Dashboard },
     {
         path: "/new-opportunity",
         exact: true,
@@ -62,6 +60,6 @@ export const routes = [
     {
         path: "/organization-dashboard",
         exact: true,
-        component: OrganizationDashboard,
+        component: Dashboard,
     },
 ];
